Missing a size_t definition that causes yet more segfaulting...
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@966 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Cubes/drawMomentCutout.cc
@@ -92,7 +92,7 @@ namespace duchamp
       float *image = new float[size*size];
       for(size_t i=0;i<size*size;i++) image[i]=0.;
 
-      int imPos,cubePos;
+      size_t imPos,cubePos;
       for(int z=zmin; z<=zmax; z++){
 	for(int x=xmin; x<=xmax; x++){
 	  for(int y=ymin; y<=ymax; y++){
